VAMO

+++ ft_strlen.c
@@ -6,16 +6,16 @@
 /*   By: rdel-fra <rdel-fra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/07 15:43:54 by rdel-fra          #+#    #+#             */
-/*   Updated: 2024/10/10 13:19:41 by rdel-fra         ###   ########.fr       */
+/*   Updated: 2024/10/11 18:22:50 by rdel-fra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <string.h>
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
